async_serial_comm/async_serial_py/pyproject.toml
[project]
name = "async-serial-py"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Reto Trappitsch", email = "reto@galactic-forensics.space" }
]
dependencies = [
    "pyserial>=3.5",
    "pyserial-asyncio>=0.6",
    "aioserial>=1.3.1",
]
readme = "README.md"
requires-python = ">= 3.11"
license = { text = "MIT" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.rye.scripts]
test = "python -m async_serial_py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/async_serial_py"]
